Prevent APC causing random fatal errors [fixes #264]
Update the phar stub to prevent APC attempting to run the composer classes through
the opcode cache on supported APC versions to avoid the random fatal "duplicate
class" errors at runtime as reported in #264.

Prior to APC version 3.0.12 this was a PHP_INI_SYSTEM directive, so output a
warning instead.