Check headers for missing scopes before asking for new OAuth token
This patch stops the GitHub VCS driver prompting for a new access
token when a repository is deleted/hidden.

Specifically, it checks the X-OAuth-Scopes and X-Accepted-OAuth-Scopes
response headers to see if the scopes on the current request match
those needed by the API call. If they do, the 404 means that the repo
is deleted/hidden, and there's no point asking for a new OAuth token.