Fix case insensitivity of ‘require’ command
When currently executing the `require` command for a package that is already listed in `require(-dev)`, one must use the exact same, case matching package name as written in `composer.json`. That is, if one changes the case of a character in the package name, the `require` command will add a new entry to `require(-dev)`, instead of updating the existing one.

This commit fixes the described behaviour to make it consistent with other commands like `update` that are already case insensitive.