Allow Composer to be used without running the application
For [Humbug
Box](https://github.com/humbug/box/blob/master/src/Composer/ComposerOrchestrator.php#L30) we are
using Composer to dump the autoload. To do so I'm using the `Composer` class from the application:

```php
$composer = (new ComposerApplication())->getComposer();
```

If you do so however this is going to fail because `Application#io` is null instead of being a
`IOInterface` instance. Indeed it is initialised only when the application is run. So one solution
is to initialised it with a dummy IO and the right IO object will be set when the application is run
as usual.