Fix running tests with non standard default branch
When the global git config has init.defaultbranch set to
something other than master the tests fail. This is because
the tests assume that the default branch is called master
but that may not be the case.

This fixes the tests by ensuring that when running git init
we ensure the default branch is called master regardless of
how git is configured.